requirements.txt
pdfplumber

python-docx
spacy
scikit-learn
PyYAML
rich
tqdm
streamlit